Corrigindo as dimensões da tela 'about' em dispositivos móveis e definindo dimensões para o card 'about'
