+++ src/app/(site)/about/components/card/about-card.types.ts
