+++ src/app/(site)/about/page.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import LoaggedUser from "@/app/ui/layout/logged-user/logged-user";
 import { aboutMock, cardTexts } from "./mock/about.mock";
-import AboutCard from "./layout/card/about-card.component";
+import AboutCard from "./components/card/about-card.component";
 
 function About() {
   return (
@@ -13,15 +13,19 @@ function About() {
         flex-col
         items-center
         justify-start
-        gap-16 
-        h-full
-        animate-[fadeIn_0.3s_ease-in-out_forwards]
-        w-3/6
+        w-screen
         shadow-md
-        bg-background
         overflow-scroll
         overflow-x-hidden
+        min-h-screen
+        lg:w-3/6
+        h-full
+        bg-background
+        gap-16 
+        py-10
+        animate-[fadeIn_0.3s_ease-in-out_forwards]
         "
+        // animate-[fadeIn_0.3s_ease-in-out_forwards]
       >
         <div className="flex flex-col gap-6 p-7">
           <h1 className="text-center text-3xl mb-4 font-black text-gray">
@@ -37,22 +41,18 @@ function About() {
           <h2 className="uppercase font-black text-gray">
             objetivos do projeto
           </h2>
-          <div className="grid grid-cols-2 items-center justify-start gap-10">
-            <AboutCard
-              image="https://www.integrasist.com.br/dados/editor/image/WhatsApp_Image_2021_04_05_at_21.13.13.jpeg"
-              title="De desenvolvedores para desenvolvedores"
-              text="Nossa equipe é formada por desenvolvedores apaixonados que vivem e respiram código. Estamos imersos no mundo da programação e entendemos as necessidades e desejos dos nossos colegas desenvolvedores.         Nossa plataforma é um espaço onde capacita a comunidade de desenvolvedores a criar um ambiente que inspire criatividade, aprendizado e inovação."
-            />
-            <AboutCard
-              image="https://www.integrasist.com.br/dados/editor/image/WhatsApp_Image_2021_04_05_at_21.13.13.jpeg"
-              title="De desenvolvedores para desenvolvedores"
-              text="Nossa equipe é formada por desenvolvedores apaixonados que vivem e respiram código. Estamos imersos no mundo da programação e entendemos as necessidades e desejos dos nossos colegas desenvolvedores.         Nossa plataforma é um espaço onde capacita a comunidade de desenvolvedores a criar um ambiente que inspire criatividade, aprendizado e inovação."
-            />
-            <AboutCard
-              image="https://www.integrasist.com.br/dados/editor/image/WhatsApp_Image_2021_04_05_at_21.13.13.jpeg"
-              title="De desenvolvedores para desenvolvedores"
-              text="Nossa equipe é formada por desenvolvedores apaixonados que vivem e respiram código. Estamos imersos no mundo da programação e entendemos as necessidades e desejos dos nossos colegas desenvolvedores.         Nossa plataforma é um espaço onde capacita a comunidade de desenvolvedores a criar um ambiente que inspire criatividade, aprendizado e inovação."
-            />
+          <div className="grid grid-cols-1 xl:grid-cols-2 gap-10">
+            {
+              aboutMock.map((card, index) => (
+                <div key={index}>
+                  <AboutCard
+                    title={card.title}
+                    text={card.text}
+                    image={card.image}
+                  />
+                </div>
+              ))
+            }
           </div>
         </div>
       </div>
